pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "activity-monitor"
version = "0.1.0"
description = "An LLM Agent to monitor system activity"
authors = ["Abishek <abishek@thoughtworks.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "activitymonitor"}]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.130.0"
crewai-tools = "^0.47.1"
python-dotenv = "^1.1.1"
openai = "^1.60.1"
langchain = "^0.3.26"
langchain-openai = "^0.2.6"
psutil = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
types-psutil = "^7.0.0"
